+++ include/nbl/asset/utils/IShaderCompiler.h
@@ -178,7 +178,7 @@ class NBL_API2 IShaderCompiler : public core::IReferenceCounted
 
 		@returns Shader containing logically same High Level code as input but with #include directives resolved.
 		*/
-		virtual std::string preprocessShader(std::string&& code, IShader::E_SHADER_STAGE& stage, const SPreprocessorOptions& preprocessOptions) const;
+		virtual std::string preprocessShader(std::string&& code, IShader::E_SHADER_STAGE& stage, const SPreprocessorOptions& preprocessOptions) const = 0;
 
 		std::string preprocessShader(system::IFile* sourcefile, IShader::E_SHADER_STAGE stage, const SPreprocessorOptions& preprocessOptions) const;
 		
@@ -260,76 +260,6 @@ class NBL_API2 IShaderCompiler : public core::IReferenceCounted
 		CIncludeFinder* getDefaultIncludeFinder() { return m_defaultIncludeFinder.get(); }
 
 		const CIncludeFinder* getDefaultIncludeFinder() const { return m_defaultIncludeFinder.get(); }
-
-		//string to be replaced with all "#" except those in "#include"
-		static constexpr const char* PREPROC_DIRECTIVE_DISABLER = "_this_is_a_hash_";
-		static constexpr const char* PREPROC_DIRECTIVE_ENABLER = PREPROC_DIRECTIVE_DISABLER;
-		static constexpr const char* PREPROC_GL__DISABLER = "_this_is_a_GL__prefix_";
-		static constexpr const char* PREPROC_GL__ENABLER = PREPROC_GL__DISABLER;
-		static constexpr const char* PREPROC_LINE_CONTINUATION_DISABLER = "_this_is_a_line_continuation_\n";
-		static constexpr const char* PREPROC_LINE_CONTINUATION_ENABLER = "_this_is_a_line_continuation_";
-
-		static void disableAllDirectivesExceptIncludes(std::string& _glslCode)
-		{
-			// TODO: replace this with a proper-ish proprocessor and includer one day
-			std::regex directive("#(?!(include|version|pragma shader_stage|line))");//all # not followed by "include" nor "version" nor "pragma shader_stage"
-			//`#pragma shader_stage(...)` is needed for determining shader stage when `_stage` param of IShaderCompiler functions is set to ESS_UNKNOWN
-			auto result = std::regex_replace(_glslCode, directive, PREPROC_DIRECTIVE_DISABLER);
-			std::regex glMacro("[ \t\r\n\v\f]GL_");
-			result = std::regex_replace(result, glMacro, PREPROC_GL__DISABLER);
-			std::regex lineContinuation("\\\\[ \t\r\n\v\f]*\n");
-			_glslCode = std::regex_replace(result, lineContinuation, PREPROC_LINE_CONTINUATION_DISABLER);
-		}
-		static void reenableDirectives(std::string& _glslCode)
-		{
-			std::regex lineContinuation(PREPROC_LINE_CONTINUATION_ENABLER);
-			auto result = std::regex_replace(_glslCode, lineContinuation, " \\");
-			std::regex glMacro(PREPROC_GL__ENABLER);
-			result = std::regex_replace(result, glMacro, " GL_");
-			std::regex directive(PREPROC_DIRECTIVE_ENABLER);
-			_glslCode = std::regex_replace(result, directive, "#");
-		}
-		static std::string encloseWithinExtraInclGuards(std::string&& _glslCode, uint32_t _maxInclusions, const char* _identifier)
-		{
-			assert(_maxInclusions != 0u);
-
-			using namespace std::string_literals;
-			std::string defBase_ = "_GENERATED_INCLUDE_GUARD_"s + _identifier + "_";
-			std::replace_if(defBase_.begin(), defBase_.end(), [](char c) ->bool { return !::isalpha(c) && !::isdigit(c); }, '_');
-
-			auto genDefs = [&defBase_, _maxInclusions, _identifier] {
-				auto defBase = [&defBase_](uint32_t n) { return defBase_ + std::to_string(n); };
-				std::string defs = "#ifndef " + defBase(0) + "\n\t#define " + defBase(0) + "\n";
-				for (uint32_t i = 1u; i <= _maxInclusions; ++i) {
-					const std::string defname = defBase(i);
-					defs += "#elif !defined(" + defname + ")\n\t#define " + defname + "\n";
-				}
-				defs += "#endif\n";
-				return defs;
-			};
-			auto genUndefs = [&defBase_, _maxInclusions, _identifier] {
-				auto defBase = [&defBase_](int32_t n) { return defBase_ + std::to_string(n); };
-				std::string undefs = "#ifdef " + defBase(_maxInclusions) + "\n\t#undef " + defBase(_maxInclusions) + "\n";
-				for (int32_t i = _maxInclusions - 1; i >= 0; --i) {
-					const std::string defname = defBase(i);
-					undefs += "#elif defined(" + defname + ")\n\t#undef " + defname + "\n";
-				}
-				undefs += "#endif\n";
-				return undefs;
-			};
-
-			return
-				genDefs() +
-				"\n"
-				"#ifndef " + defBase_ + std::to_string(_maxInclusions) +
-				"\n" +
-				"#line 1 \"" + _identifier + "\"\n" +
-				_glslCode +
-				"\n"
-				"#endif"
-				"\n\n" +
-				genUndefs();
-		}
 	protected:
 
 		virtual void insertIntoStart(std::string& code, std::ostringstream&& ins) const = 0;
